+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 ## 0.0.2
 
 - add `getNearbyCityByLatLng` functionality
+
+## 0.0.3
+
+- change Korean key names to English
+- change distance param type from int to double
+- [bug fix] replace lat and lng